+++ CHANGELOG.rst
@@ -4,6 +4,13 @@ Change Log
 
 HEAD
 ----
+* Added ``cancel`` to ``CalendarItem`` and ``CancelCalendarItem`` class to allow cancelling meetings that were set up
+* Added ``accept``, ``decline`` and ``tentatively_accept`` to ``CalendarItem`` as wrapper methods
+* Added ``accept``, ``decline`` and ``tentatively_accept`` to ``MeetingRequest`` to respond to incoming invitations
+* Added ``BaseMeetingItem`` (inheriting from ``Item``) being used as base for MeetingCancellation, MeetingMessage,
+  MeetingRequest and MeetingResponse
+* Added ``AssociatedCalendarItemId`` (property), ``AssociatedCalendarItemIdField`` and ``ReferenceItemIdField``
+* Added ``PostReplyItem``
 * Removed ``Folder.get_folder_by_name()`` which has been deprecated since version ``1.10.2``.
 * Added ``Item.copy(to_folder=some_folder)`` method which copies an item to the given folder and
   returns the ID of the new item.